pyproject.toml
[tool.poetry]
name = "yaeb"
version = "1.0.0"
description = "A simple typed event bus written in pure python"
authors = ["Daniil Fedyaev <wintercitizen@outlook.com>"]
readme = "README.md"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
repository = "https://github.com/WinterCitizen/yaeb"
keywords = ["event", "bus", "typed", "eventbus"]
include = ["yaeb/py.typed"]

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
pytest = "*"
mypy = "*"
black = "*"
pytest-sugar = "*"
ipdb = "*"
pytest-cov = "*"
wemake-python-styleguide = "*"
pycln = "*"
isort = "*"
add-trailing-comma = "*"
pytest-asyncio = "*"
factory-boy = "*"

[tool.coverage.report]
exclude_lines = [
    "raise NotImplementedError",
    "if TYPE_CHECKING",
]

[tool.isort]
multi_line_output = "VERTICAL_HANGING_INDENT"
include_trailing_comma = true

[tool.mypy]
disallow_subclassing_any = false
ignore_missing_imports = true
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
